remove the deprecated 'upload' screen, navigate directly to preview
