+++ lib/utils/constants.dart
@@ -29,7 +29,6 @@ const kIconDenomEmoney = 'assets/images/denom_emoney.png';
 const kIconDenomAgoric = 'assets/images/denom_agoric.png';
 const kIconDenomLuna = 'assets/images/denom_luna.png';
 const kIconDenomJuno = 'assets/images/denom_juno.png';
-
 const kTextFieldSingleLine = 'assets/images/text_field_single_line.png';
 const kTextFieldMultiLine = 'assets/images/text_field_multi_line.png';
 const kTextFieldButton = 'assets/images/text_field_button.png';
@@ -134,12 +133,12 @@ const kSizeText = 'Size';
 const kDurationText = 'Duration';
 const kDateText = 'Date';
 const kRoyaltyText = 'Royalty';
-const kPreviewNFTText = 'Preview your NFT';
+const kPreviewNFTText = '';
 const kPreview3dModelText = 'Click here to preview\nyour selected 3D Model';
 const kMintMoreText = 'Mint More';
 const kGoToWalletText = 'Go to Wallet';
 const kChooseNFTFormatText = 'Choose your NFT format';
-const kUploadNFTText = 'Upload your NFT file';
+const kUploadNFTText = 'Upload NFT file';
 const kEditNFTText = 'Edit your NFT Details';
 const kUploadText = 'Upload';
 const kPreviewText = 'Preview';
@@ -165,7 +164,7 @@ const kTapToSelect = 'Tap to Select';
 const kCloseText = 'Close';
 const kUploadHint2 = '• Image, Video, 3D or Audio';
 const kUploadHint3 = '• One file per upload';
-const kUploadHintAll = 'GB Limit\nOne file per upload';
+const kUploadHintAll = 'GB Limit.\nOne file per upload.';
 const kHintNftName = 'Bird on Shoulder';
 const kHintArtistName = 'Sarah Jackson';
 const kHintNftDesc =
